+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # v0.3.1
 
-- Added `Reader.readPath()` for fast path-based queries on encoded data using Javascript-style property access: `"name"`, `"address.city"`, `"items[0]"`, `"users[5].scores[0]"`.
+- Added `Reader.readPath` for fast path-based queries on encoded data using Javascript-style property access: `"name"`, `"address.city"`, `"items[0]"`, `"users[5].scores[0]"`.
+- Added `Reader.readPaths` for efficiently reading a list of paths in a single pass.
 - Added `Writer.applyUpdates` for efficiently applying a changeset of updates to an existing encoded buffer in a single pass.
 
 ### Improvements